Revamp the offline player access to work with 1.15.2.
My players on my server had a problem, their UUIDs changed, and they lost access to
their inventories.  So I had to revamp this plugin.  It does NOT remove the items
from the save files, in this version, so it shouldn't be exposed to users because it
dupes items, essentially.